+++ models/CSFI2.py
@@ -1,0 +1,26 @@
+from models.HyperTransformer_V2 import conv1x1, conv3x3
+
+
+import torch
+import torch.nn.functional as F
+from torch import nn
+
+
+class CSFI2(nn.Module):
+    def __init__(self, n_feats):
+        super(CSFI2, self).__init__()
+        self.conv12 = conv1x1(n_feats, int(n_feats/2))
+        self.conv21 = conv3x3(int(n_feats/2), n_feats, 2)
+
+        self.conv_merge1 = conv3x3(n_feats*2, n_feats)
+        self.conv_merge2 = conv3x3(n_feats, int(n_feats/2))
+
+    def forward(self, x1, x2):
+        x12 = F.interpolate(x1, scale_factor=2, mode='bicubic')
+        x12 = F.relu(self.conv12(x12))
+        x21 = F.relu(self.conv21(x2))
+
+        x1 = F.relu(self.conv_merge1( torch.cat((x1, x21), dim=1) ))
+        x2 = F.relu(self.conv_merge2( torch.cat((x2, x12), dim=1) ))
+
+        return x1, x2
