+++ models/NoAttention.py
@@ -1,0 +1,12 @@
+from torch import nn
+
+
+class NoAttention(nn.Module):
+    ''' Scaled Dot-Product Attention '''
+
+    def __init__(self):
+        super().__init__()
+
+    def forward(self, v, k, q, mask=None):
+        output = v
+        return output
