+++ models/ResBlock.py
@@ -1,0 +1,21 @@
+from models.HyperTransformer_V2 import conv3x3
+
+
+from torch import nn
+
+
+class ResBlock(nn.Module):
+    def __init__(self, in_channels, out_channels, stride=1, downsample=None, res_scale=1):
+        super(ResBlock, self).__init__()
+        self.res_scale = res_scale
+        self.conv1 = conv3x3(in_channels, out_channels, stride)
+        self.relu = nn.ReLU(inplace=True)
+        self.conv2 = conv3x3(out_channels, out_channels)
+
+    def forward(self, x):
+        x1 = x
+        out = self.conv1(x)
+        out = self.relu(out)
+        out = self.conv2(out)
+        out = out * self.res_scale + x1
+        return out
