+++ models/SFE.py
@@ -1,0 +1,28 @@
+from models.HyperTransformer_V2 import conv3x3
+from models.ResBlock import ResBlock
+
+import torch.nn.functional as F
+from torch import nn
+
+
+class SFE(nn.Module):
+    def __init__(self, in_feats, num_res_blocks, n_feats, res_scale):
+        super(SFE, self).__init__()
+        self.num_res_blocks = num_res_blocks
+        self.conv_head = conv3x3(in_feats, n_feats)
+
+        self.RBs = nn.ModuleList()
+        for i in range(self.num_res_blocks):
+            self.RBs.append(ResBlock(in_channels=n_feats, out_channels=n_feats,
+                res_scale=res_scale))
+
+        self.conv_tail = conv3x3(n_feats, n_feats)
+
+    def forward(self, x):
+        x = F.relu(self.conv_head(x))
+        x1 = x
+        for i in range(self.num_res_blocks):
+            x = self.RBs[i](x)
+        x = self.conv_tail(x)
+        x = x + x1
+        return x
