+++ models/ScaledDotProductAttention.py
@@ -1,0 +1,25 @@
+import torch
+import torch.nn.functional as F
+from torch import nn
+
+#This function implements the multi-head attention
+class ScaledDotProductAttention(nn.Module):
+    ''' Scaled Dot-Product Attention '''
+
+    def __init__(self, temperature):
+        super().__init__()
+        self.temperature = temperature
+
+    def forward(self, v, k, q, mask=None):
+        # Compute attention
+        attn = torch.matmul(q / self.temperature, k.transpose(-2, -1))
+
+        if mask is not None:
+            attn = attn.masked_fill(mask == 0, -1e9)
+
+        # Normalization (SoftMax)
+        attn    = F.softmax(attn, dim=-1)
+
+        # Attention output
+        output  = torch.matmul(attn, v)
+        return output
